move simple modal to its own file hook up edit and delete ape key to modals
